atualizado o select para pegar o maior arquivo

+++ codificar.py
@@ -53,27 +53,32 @@ def codificacaoDeFilmes_Locais() -> None:
             for tipo in [arqMkv,arqMp4,arqAvi]:
                 for arq in tipo:
                     [pasta_filme,nome,arquivo,imdb,tmpPasta,tmpArq,localDoArquivo] = correcaoFilmes(arq);
-                    if(os.path.getsize(f'{localDoArquivo}') >= TAMANHO_MINIMO):
-                        os.system(f'mkdir {tmpPasta}');
-                        if(CODIFICADOR == 'HEVC'):
-                            linhaDeComando = f'ffmpeg -i "{localDoArquivo}" -c:v hevc_nvenc -c:a copy "{tmpArq}"';
-                        elif(CODIFICADOR == 'H265'):
-                            linhaDeComando = f'ffmpeg -i "{localDoArquivo}" -c:v libx265 -c:a copy "{tmpArq}"';
-                        else:
-                            linhaDeComando = f'ffmpeg -i "{localDoArquivo}" -vcodec h264 -acodec mp3 "{tmpArq}"';
-                        print(linhaDeComando);
-                        os.system(linhaDeComando);
-                    if(os.path.isfile(tmpArq)):
-                        if((os.path.getsize(tmpArq) <= os.path.getsize(localDoArquivo)) and (os.path.getsize(tmpArq) > TAMANHO_MINIMO)):
-                            os.system(f'mv {tmpArq} "{pasta_filme}"');
-                    if(os.path.exists(f'"{tmpPasta}"')):
-                        os.system(f'rm -r "{tmpPasta}"');
+                    if(os.path.isfile(localDoArquivo)):
+                        comando = f'UPDATE filmes SET codificado=2 WHERE imdb="{imdb}";';
+                        bd.update(comando=comando,cnxn=cnxn,cursor=cursor);
+                        linhaDeComando = '';
+                        if(os.path.getsize(f'{localDoArquivo}') >= TAMANHO_MINIMO):
+                            os.system(f'mkdir "{tmpPasta}"');
+                            if(CODIFICADOR == 'HEVC'):
+                                linhaDeComando = f'ffmpeg -i "{localDoArquivo}" -c:v hevc_nvenc -c:a copy "{tmpArq}"';
+                            elif(CODIFICADOR == 'H265'):
+                                linhaDeComando = f'ffmpeg -i "{localDoArquivo}" -c:v libx265 -c:a copy "{tmpArq}"';
+                            else:
+                                linhaDeComando = f'ffmpeg -i "{localDoArquivo}" -vcodec h264 -acodec mp3 "{tmpArq}"';
+                            os.system("clear");
+                            print(linhaDeComando);
+                            os.system(linhaDeComando);
+                        if(os.path.isfile(tmpArq)):
+                            if((os.path.getsize(tmpArq) <= os.path.getsize(localDoArquivo)) and (os.path.getsize(tmpArq) > TAMANHO_MINIMO)):
+                                os.system(f'mv {tmpArq} "{pasta_filme}"');
+                        if(os.path.exists(f'"{tmpPasta}"')):
+                            os.system(f'rm -r "{tmpPasta}"');
                     escrever.writerow(['Nome','Arquivo','imdb','Pasta','Tamanho']);
                     linha = [nome,arquivo,imdb,pasta_filme,str(os.path.getsize(localDoArquivo))];
                     escrever.writerow(linha);
 
 def codificacaoDeFilmes_Banco(cursor,cnxn) -> None:
-    comando = f"SELECT pasta, arquivo,imdb FROM filmes WHERE  codificado = 0 and tamanho != 0 ORDER BY pasta LIMIT 1;";
+    comando = f"SELECT pasta, arquivo,imdb FROM filmes WHERE  codificado = 0 and tamanho != 0 ORDER BY tamanho DESC LIMIT 1;";
     retorno = bd.read(comando=comando,cursor=cursor);
     while retorno != []:
         retorno                 = retorno[0];
@@ -85,7 +90,6 @@ def codificacaoDeFilmes_Banco(cursor,cnxn) -> None:
         print(f'tmp pasta = {pasta}tmp/')
         print(f'local do arquivo = "{pasta}{arquivo}"')
         print(f'tmp arquivo = "{tmpPasta}{arquivo}"')
-
         if(os.path.isfile(localDoArquivo)):
             comando = f'UPDATE filmes SET codificado=2 WHERE imdb="{imdb}";';
             bd.update(comando=comando,cnxn=cnxn,cursor=cursor);
@@ -108,7 +112,7 @@ def codificacaoDeFilmes_Banco(cursor,cnxn) -> None:
                 os.system(f'rm -r "{tmpPasta}"');
             comando = f'UPDATE filmes.filmes SET tamanho="{os.path.getsize(localDoArquivo)}", codificado=1 WHERE imdb="{imdb}";';
             bd.update(comando=comando,cnxn=cnxn,cursor=cursor);
-        comando = f"SELECT pasta, arquivo,imdb FROM filmes WHERE  codificado = 0 and tamanho != 0 ORDER BY pasta LIMIT 1;";
+        comando = f"SELECT pasta, arquivo,imdb FROM filmes WHERE  codificado = 0 and tamanho != 0 ORDER BY tamanho DESC LIMIT 1;";
         retorno = bd.read(comando=comando,cursor=cursor);
 #-----------------------
 # Main()
